+++ Sources/MapboxNavigationCore/Settings/TelemetryAppMetadata.swift
@@ -7,17 +7,30 @@ public struct TelemetryAppMetadata: Equatable, Sendable {
     public let name: String
     /// Version of the application.
     public let version: String
+
     /// User ID relevant for the application context.
+    @available(*, deprecated, message: "This property no longer has any effect.")
     public var userId: String?
     /// Session ID relevant for the application context.
+    @available(*, deprecated, message: "This property no longer has any effect.")
     public var sessionId: String?
 
-    /// nitializes a new `TelemetryAppMetadata` object.
+    /// Initializes a new ``TelemetryAppMetadata`` object.
+    /// - Parameters:
+    ///   - name: Name of the application.
+    ///   - version: Version of the application.
+    public init(name: String, version: String) {
+        self.name = name
+        self.version = version
+    }
+
+    /// Initializes a new ``TelemetryAppMetadata`` object.
     /// - Parameters:
     ///   - name: Name of the application.
     ///   - version: Version of the application.
     ///   - userId: User ID relevant for the application context.
     ///   - sessionId: Session ID relevant for the application context.
+    @available(*, deprecated, message: "Use 'init(name:version:)' instead.")
     public init(
         name: String,
         version: String,
@@ -33,16 +46,9 @@ public struct TelemetryAppMetadata: Equatable, Sendable {
 
 extension TelemetryAppMetadata {
     var configuration: [String: String?] {
-        var dictionary: [String: String?] = [
+        [
             "name": name,
             "version": version,
         ]
-        if let userId, !userId.isEmpty {
-            dictionary["userId"] = userId
-        }
-        if let sessionId, !sessionId.isEmpty {
-            dictionary["sessionId"] = sessionId
-        }
-        return dictionary
     }
 }